_putnbr

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int nbr)
 {
-	unsigned int i;
+	long int i;
 
 	i = 10;
 	if (nbr < 0)
